De-duplicate resource interface tests
* Recent tests duplicated validation tests between resource_interface
  and shader_io
* Moved resource interface testing into shader_io and de-duplicated
  functionality
* Expanded set of resources tests for group and binding requirements
